webpage/chatbot-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
openai==1.59.3
google-cloud-speech==2.28.0
google-cloud-texttospeech==2.21.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.1
